Create README.md
This game is designed based on java language

When the jar file is ran, two windows are opened the smallest window is the initializer. In this window, the user can choose the number of players (between two to four) and level number (between one to three) when he clicks on the ok button, the game starts. The first player is always is human. The playing board is located in the second window (biggest window). At the beginning of the game, you can see two cells are crossed and some cells are highlighted. The player at first can lay the die around the crossed cell, if he has the die value in his dice list then the available

cells are highlighted and he can lays his dices in these cells and after finishing the turn the filled positions are moved to the crossed list. When he laid the first die in his board, that die is removed from the dice list and next player is called. In the next turn, player can lay his die near the crossed positions if he has related die value in his dice list. If the player is not able to play then for human player two buttons “Roll” and “Drop out” is activated and he can click one of them(Figure8). If he clicks on the roll the dice list again rolled and player has another opportunity to play. If he again cannot play after rolling, he should as penalty return one die to the dice list and next player is called. If he instead of “Roll” button clicks on the drop out he skip his round and next player is called. When a player cannot play and drop out his turn the big pen is displayed in his board. These process continues until one of them win the game.